fix: Adapt lcs-leadmultispace to tabstop/shiftwidth changes
Because vim-sleuth will dynamically change the values of `tabstop` and
`shiftwidth` based on context cues and config files, a single value for
`lcs-multispace` leads to unexpected behavior where multiple vertical
lines may be shown for each level of indentation.

This implements autocmds to adjust the value of `lcs-multispace`
whenever `shiftwidth` and/or `tabstop` change, so that there will always
be a single vertical line per level of indentation, as detected by
vim-sleuth.